+++ Tests/Unit/Hooks/HookProviderTest.php
@@ -9,13 +9,10 @@ use OliverKlee\Seminars\Hooks\Interfaces\Hook;
 use OliverKlee\Seminars\Tests\Unit\Hooks\Fixtures\TestingHookImplementor;
 use OliverKlee\Seminars\Tests\Unit\Hooks\Fixtures\TestingHookImplementor2;
 use OliverKlee\Seminars\Tests\Unit\Hooks\Fixtures\TestingHookImplementorNotImplementsInterface;
-use OliverKlee\Seminars\Tests\Unit\Hooks\Fixtures\TestingHookImplementorReturnsArray;
-use OliverKlee\Seminars\Tests\Unit\Hooks\Fixtures\TestingHookImplementorReturnsArray2;
 use OliverKlee\Seminars\Tests\Unit\Hooks\Fixtures\TestingHookImplementorReturnsModifiedValue;
 use OliverKlee\Seminars\Tests\Unit\Hooks\Fixtures\TestingHookImplementorReturnsModifiedValue2;
 use OliverKlee\Seminars\Tests\Unit\Hooks\Fixtures\TestingHookInterface;
 use OliverKlee\Seminars\Tests\Unit\Hooks\Fixtures\TestingHookInterfaceNotExtendsHook;
-use OliverKlee\Seminars\Tests\Unit\Hooks\Fixtures\TestingHookInterfaceReturnsArray;
 use OliverKlee\Seminars\Tests\Unit\Hooks\Fixtures\TestingHookInterfaceReturnsModifiedValue;
 use TYPO3\TestingFramework\Core\Unit\UnitTestCase;
 
@@ -348,106 +345,6 @@ final class HookProviderTest extends UnitTestCase
 
         self::assertSame(0, TestingHookImplementor::$wasCalled);
         self::assertSame(1, TestingHookImplementor2::$wasCalled);
-    }
-
-    // Tests concerning Hook::executeHookReturningMergedArray().
-
-    /**
-     * @test
-     */
-    public function hookObjectForTestingHookReturningArrayCanBeCreated(): void
-    {
-        self::assertInstanceOf(HookProvider::class, new HookProvider(TestingHookInterfaceReturnsArray::class));
-    }
-
-    /**
-     * @test
-     */
-    public function hookObjectForTestingHookReturningArrayWithNoHookImplementorRegisteredReturnsEmptyArray(): void
-    {
-        $hookObject = new HookProvider(TestingHookInterfaceReturnsArray::class);
-
-        self::assertSame([], $hookObject->executeHookReturningMergedArray('testHookMethodReturnsArray'));
-        self::assertSame([], $hookObject->executeHookReturningMergedArray('testHookMethodReturnsNestedArray'));
-    }
-
-    /**
-     * @test
-     */
-    public function hookObjectForTestingHookReturningArrayWithNoHookImplementorRegisteredFailsForEmptyMethod(): void
-    {
-        $hookObject = new HookProvider(TestingHookInterfaceReturnsArray::class);
-
-        $this->expectException(\InvalidArgumentException::class);
-        $this->expectExceptionCode(1573479911);
-
-        $hookObject->executeHookReturningMergedArray('');
-    }
-
-    /**
-     * @test
-     */
-    public function hookObjectForTestingHookReturningArrayWithNoHookImplementorRegisteredFailsForUnknownMethod(): void
-    {
-        $hookObject = new HookProvider(TestingHookInterfaceReturnsArray::class);
-
-        $this->expectException(\UnexpectedValueException::class);
-        $this->expectExceptionCode(1573480302);
-
-        $hookObject->executeHookReturningMergedArray('methodNotImplemented');
-    }
-
-    /**
-     * @test
-     */
-    public function hookObjectForTestingHookReturningArrayWithOneHookImplementorRegisteredReturnsFilledArray(): void
-    {
-        $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['seminars'][TestingHookInterfaceReturnsArray::class][1577363366] =
-            TestingHookImplementorReturnsArray::class;
-        $hookObject = new HookProvider(TestingHookInterfaceReturnsArray::class);
-
-        self::assertSame(
-            ['me' => 'ok', 'overwritten' => 'initial'],
-            $hookObject->executeHookReturningMergedArray('testHookMethodReturnsArray'),
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function hookObjectForTestingHookReturningArrayWithTwoHookImplementorsRegisteredReturnsMergedArray(): void
-    {
-        $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['seminars'][TestingHookInterfaceReturnsArray::class][1577363366] =
-            TestingHookImplementorReturnsArray::class;
-        $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['seminars'][TestingHookInterfaceReturnsArray::class][1577363367] =
-            TestingHookImplementorReturnsArray2::class;
-        $hookObject = new HookProvider(TestingHookInterfaceReturnsArray::class);
-
-        self::assertSame(
-            ['me' => 'ok', 'overwritten' => 'replaced', 'me2' => 'ok'],
-            $hookObject->executeHookReturningMergedArray('testHookMethodReturnsArray'),
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function hookObjectForTestingHookReturningArrayWithTwoHookImplementorsRegisteredReturnsNestedMergedArray(
-    ): void {
-        $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['seminars'][TestingHookInterfaceReturnsArray::class][1577363366] =
-            TestingHookImplementorReturnsArray::class;
-        $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['seminars'][TestingHookInterfaceReturnsArray::class][1577363367] =
-            TestingHookImplementorReturnsArray2::class;
-        $hookObject = new HookProvider(TestingHookInterfaceReturnsArray::class);
-
-        self::assertSame(
-            [
-                'me' => ['status' => true],
-                'me1' => ['status' => true],
-                'me2' => ['status' => false, 'newValue' => 'new2'],
-            ],
-            $hookObject->executeHookReturningMergedArray('testHookMethodReturnsNestedArray'),
-        );
     }
 
     // Tests concerning Hook::executeHookReturningModifiedValue().